+++ OCMS/OCMS.csproj
@@ -47,18 +47,28 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Bussiness\MedicineBusiness.cs" />
+    <Compile Include="Bussiness\UsersBusiness.cs" />
     <Compile Include="Class\clsMain.cs" />
     <Compile Include="Bussiness\ConsultationBusiness.cs" />
     <Compile Include="Bussiness\MemberBussiness.cs" />
+    <Compile Include="Class\clsUtility.cs" />
     <Compile Include="DAL\ConsultationDataAccess.cs" />
     <Compile Include="DAL\DataAccess.cs" />
     <Compile Include="DAL\MedicineDataAccess.cs" />
     <Compile Include="DAL\MemberDataAccess.cs" />
+    <Compile Include="DAL\UsersDataAccess.cs" />
     <Compile Include="DBResource.cs" />
     <Compile Include="MODEL\DiagnosisModel.cs" />
     <Compile Include="MODEL\DispositionModel.cs" />
     <Compile Include="MODEL\MedicineModel.cs" />
+    <Compile Include="MODEL\UsersModel.cs" />
     <Compile Include="Utilities.cs" />
+    <Compile Include="frmConfirmationPage.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="frmConfirmationPage.Designer.cs">
+      <DependentUpon>frmConfirmationPage.cs</DependentUpon>
+    </Compile>
     <Compile Include="VIEW\frmOMList.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -100,8 +110,17 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="VIEW\frmUsersInformation.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="VIEW\frmUsersInformation.Designer.cs">
+      <DependentUpon>frmUsersInformation.cs</DependentUpon>
+    </Compile>
     <EmbeddedResource Include="frmLogin.resx">
       <DependentUpon>frmLogin.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="frmConfirmationPage.resx">
+      <DependentUpon>frmConfirmationPage.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="VIEW\frmOMList.resx">
       <DependentUpon>frmOMList.cs</DependentUpon>
@@ -124,6 +143,9 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <EmbeddedResource Include="VIEW\frmUsersInformation.resx">
+      <DependentUpon>frmUsersInformation.cs</DependentUpon>
+    </EmbeddedResource>
     <None Include="OCMSDataSet.xsc">
       <DependentUpon>OCMSDataSet.xsd</DependentUpon>
     </None>
